pyproject.toml
[project]
name = "google-calendar-mcp"
version = "0.1.0"
description = "MCP server for Google Calendar integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
]

[project.scripts]
google-calendar-mcp = "calendar_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_system"
